requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
blinker==1.8.2
Cartopy==0.24.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
discord.py==2.4.0
dnspython==2.7.0
Flask==3.0.3
fonttools==4.54.1
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
motor==3.6.0
multidict==6.1.0
numpy==2.1.2
packaging==24.1
pillow==11.0.0
propcache==0.2.0
psutil==6.1.0
pymongo==4.9.2
pyparsing==3.2.0
pyproj==3.7.0
pyshp==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
setuptools==75.3.0
shapely==2.0.6
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.0
yarl==1.17.1
